Adding word positioning information to the transformer encoder
Since the transformer is a sequence model, word positioning is necessary though it can perfectly work fine without considering it on the sentiment classification task. Word positioning can be injected in different ways, here we used position embedded vectors instead of adding a vector with values ranging [-1, 1] to the word embeddings. (from the "attention is all you need" paper)